Plain view without database access
