--- app/src/main/java/com/isis3510/growhub/model/objects/Profile.kt
@@ -1,10 +1,0 @@
-package com.isis3510.growhub.model.objects
-
-data class Profile (
-    val name: String = "",
-    val following: Int = 0,
-    val followers: Int = 0,
-    val aboutMe: String = "",
-    val interests: List<String> = listOf(),
-    val profilePictureUrl: String = ""
-)
